requirements.txt
hyperon==0.2.5
google-genai==1.18.0
python-dotenv==1.1.0